Add Infrastructure for Dynamic Connection Level Summary Vectors
This commit adds support for registering new summary vectors at the
connection level at runtime.  This is in support of fracturing
processes which dynamically expose parts of the geometry to well
flow as a fracture opens up.  We introduce the notion of "dynamic"
connections, represented as a vector of pairs of well names and
vectors of intersected cell IDs.  We expect that these will be
created by the simulator as needed and registered with the Summary
object as they become available.